Simple Two Number Calculator
A simple two number calculator that can add, subtract, multiply and divide two numbers. Includes helpful comments for those new to python and in general programming. The same code structure can be reused to create more complex calculators in python or other programming language. I would like to see others try to change around the code and make it your own or make an entirely new program or concept out of. Every programmer has to start somewhere, don't rush to complex topics, start from the bottom and keep on learning and you'll make it to the top.